+++ lib/src/platform_search_bar.dart
@@ -72,6 +72,7 @@ class MaterialSearchBarData extends _BaseData {
     this.contextMenuBuilder,
     this.onTapOutside,
     this.scrollPadding,
+    this.readOnly,
   });
 
   // final String? hintText;
@@ -95,6 +96,7 @@ class MaterialSearchBarData extends _BaseData {
   final EditableTextContextMenuBuilder? contextMenuBuilder;
   final TapRegionCallback? onTapOutside;
   final EdgeInsets? scrollPadding;
+  final bool? readOnly;
 }
 
 class CupertinoSearchBarData extends _BaseData {
@@ -136,18 +138,18 @@ class CupertinoSearchBarData extends _BaseData {
     this.cursorOpacityAnimates,
     this.cursorRadius,
     this.cursorWidth,
-  })  : assert(
-          !((decoration != null) && (backgroundColor != null)),
-          'Cannot provide both a background color and a decoration\n'
-          'To provide both, use "decoration: BoxDecoration(color: '
-          'backgroundColor)"',
-        ),
-        assert(
-          !((decoration != null) && (borderRadius != null)),
-          'Cannot provide both a border radius and a decoration\n'
-          'To provide both, use "decoration: BoxDecoration(borderRadius: '
-          'borderRadius)"',
-        );
+  }) : assert(
+         !((decoration != null) && (backgroundColor != null)),
+         'Cannot provide both a background color and a decoration\n'
+         'To provide both, use "decoration: BoxDecoration(color: '
+         'backgroundColor)"',
+       ),
+       assert(
+         !((decoration != null) && (borderRadius != null)),
+         'Cannot provide both a border radius and a decoration\n'
+         'To provide both, use "decoration: BoxDecoration(borderRadius: '
+         'borderRadius)"',
+       );
   final ValueChanged<String>? onSubmitted;
   final BoxDecoration? decoration;
   final BorderRadius? borderRadius;
@@ -226,7 +228,8 @@ class PlatformSearchBar
     EditableTextState editableTextState,
   ) {
     return AdaptiveTextSelectionToolbar.editableText(
-        editableTextState: editableTextState);
+      editableTextState: editableTextState,
+    );
   }
 
   @override
@@ -243,16 +246,19 @@ class PlatformSearchBar
       onTap: data?.onTap ?? onTap,
       onChanged: data?.onChanged ?? onChanged,
 
-      backgroundColor: data?.backgroundColor ??
+      backgroundColor:
+          data?.backgroundColor ??
           (backgroundColor != null
               ? WidgetStateProperty.all<Color>(backgroundColor)
               : null),
       hintText: data?.hintText ?? hintText,
-      hintStyle: data?.hintStyle ??
+      hintStyle:
+          data?.hintStyle ??
           (hintStyle != null
               ? WidgetStateProperty.all<TextStyle>(hintStyle)
               : null),
-      textStyle: data?.textStyle ??
+      textStyle:
+          data?.textStyle ??
           (textStyle != null
               ? WidgetStateProperty.all<TextStyle>(textStyle)
               : null),
@@ -278,6 +284,7 @@ class PlatformSearchBar
           data?.contextMenuBuilder ?? _defaultContextMenuBuilder,
       onTapOutside: data?.onTapOutside,
       scrollPadding: data?.scrollPadding ?? const EdgeInsets.all(20.0),
+      readOnly: data?.readOnly ?? false,
     );
   }
 
@@ -306,10 +313,12 @@ class PlatformSearchBar
       borderRadius: data?.borderRadius,
       itemColor: data?.itemColor ?? CupertinoColors.secondaryLabel,
       itemSize: data?.itemSize ?? 20.0,
-      prefixInsets: data?.prefixInsets ??
+      prefixInsets:
+          data?.prefixInsets ??
           const EdgeInsetsDirectional.fromSTEB(6, 0, 0, 3),
       prefixIcon: data?.prefixIcon ?? const Icon(CupertinoIcons.search),
-      suffixInsets: data?.suffixInsets ??
+      suffixInsets:
+          data?.suffixInsets ??
           const EdgeInsetsDirectional.fromSTEB(0, 0, 5, 2),
       suffixIcon:
           data?.suffixIcon ?? const Icon(CupertinoIcons.xmark_circle_fill),